Move test from test_dataset to test_basic, move check for filesystem till after fsspec-case has been handled
