+++ python/pyarrow/parquet.py
@@ -1472,12 +1472,6 @@ def _make_manifest(path_or_paths, fs, pathsep='/', metadata_nthreads=1,
 
 
 def _is_local_file_system(fs):
-    import fsspec
-
-    if isinstance(fs, fsspec.AbstractFileSystem):
-        if type(fs).__name__ == "LocalFileSystem":
-            return True
-        return False
     return isinstance(fs, LocalFileSystem) or isinstance(
         fs, legacyfs.LocalFileSystem
     )
@@ -1503,16 +1497,6 @@ class _ParquetDatasetV2:
                     "Keyword '{0}' is not yet supported with the new "
                     "Dataset API".format(keyword))
 
-        if (
-            hasattr(path_or_paths, "__fspath__") and
-            filesystem is not None and
-            not _is_local_file_system(filesystem)
-        ):
-            raise TypeError(
-                "Path-like objects with __fspath__ must only be used with "
-                f"local file systems, not {type(filesystem)}"
-            )
-
         # map format arguments
         read_options = {}
         if buffer_size:
@@ -1533,6 +1517,18 @@ class _ParquetDatasetV2:
             # if memory_map is specified, assume local file system (string
             # path can in principle be URI for any filesystem)
             filesystem = LocalFileSystem(use_mmap=memory_map)
+
+        # This needs to be checked after _ensure_filesystem, because that handles the
+        # case of an fsspec LocalFileSystem
+        if (
+            hasattr(path_or_paths, "__fspath__") and
+            filesystem is not None and
+            not _is_local_file_system(filesystem)
+        ):
+            raise TypeError(
+                "Path-like objects with __fspath__ must only be used with "
+                f"local file systems, not {type(filesystem)}"
+            )
 
         # check for single fragment dataset
         single_file = None
